# Inline vector encoders and decoders