--- a/binary-serialise-cbor.cabal
+++ b/binary-serialise-cbor.cabal
@@ -18,7 +18,7 @@ description:
   purposes. By providing a set of type class instances, you can
   also serialize any custom data type you have as well.
   .
-  The underlying binary format used is the 'Consise Binary Object
+  The underlying binary format used is the 'Concise Binary Object
   Representation', or CBOR, specified in RFC 7049. As a result,
   serialized Haskell values have implicit structure outside of the
   Haskell program itself, meaning they can be inspected or analyzed