--- a/binary-serialise-cbor.cabal
+++ b/binary-serialise-cbor.cabal
@@ -243,6 +243,7 @@ benchmark vs-other-libs
     directory               >= 1.0     && < 1.3,
     ghc-prim                >= 0.3     && < 0.6,
     text                    >= 1.1     && < 1.3,
+    vector                  >= 0.10    && < 1.0,
     binary-serialise-cbor,
 
     filepath                >= 1.0     && < 1.5,