The `Vector` encoders and decoders should be inline lest we are forced to dynamically dispatch various calls through a ``Data.Vector.Generic.Vector` dictionary, which has a rather drastic effect on performance,

benchmark                                              |        before |                  |       after    |                  |     delta
------------------------------------------------------ |-------------- | ---------------- | -------------- | ---------------- | -----------
boxed/n=500/decode                                     |        28.23  | ±       0.8 μs   |          26.7  | ±      1.00 μs   |     -5.5%
boxed/n=500/encode                                     |        23.54  | ±       1.2 μs   |          11.5  | ±      0.41 μs   |     -51.0%
boxed/n=5000/decode                                    |       487.68  | ±      21.4 μs   |         481.4  | ±     48.74 μs   |     -1.3%
boxed/n=5000/encode                                    |       237.79  | ±      11.2 μs   |         112.6  | ±      3.90 μs   |     -52.7%
boxed/n=50000/decode                                   |      9000.91  | ±     763.0 μs   |        8472.4  | ±    352.90 μs   |     -5.9%
boxed/n=50000/encode                                   |      2573.48  | ±     321.2 μs   |        1255.5  | ±     91.90 μs   |     -51.2%
boxed/n=500000/decode                                  |    116484.39  | ±   25379.6 μs   |       98732.9  | ±   3584.60 μs   |     -15.2%
boxed/n=500000/encode                                  |     24388.84  | ±    1143.1 μs   |       12165.4  | ±    418.13 μs   |     -50.1%
unboxed/n=500/decode                                   |        47.30  | ±       4.2 μs   |          27.0  | ±      0.65 μs   |     -42.8%
unboxed/n=500/encode                                   |        33.28  | ±       5.5 μs   |          11.1  | ±      1.43 μs   |     -66.6%
unboxed/n=5000/decode                                  |       745.26  | ±      19.5 μs   |         463.9  | ±     13.53 μs   |     -37.7%
unboxed/n=5000/encode                                  |       329.08  | ±       8.0 μs   |         106.2  | ±      3.31 μs   |     -67.7%
unboxed/n=50000/decode                                 |     11070.82  | ±     343.4 μs   |        8660.3  | ±    358.24 μs   |     -21.8%
unboxed/n=50000/encode                                 |      3220.04  | ±      76.6 μs   |        1124.1  | ±     54.28 μs   |     -65.1%
unboxed/n=500000/decode                                |    115083.09  | ±    3525.2 μs   |       97234.6  | ±   3125.91 μs   |     -15.5%
unboxed/n=500000/encode                                |     32053.50  | ±     569.1 μs   |       10564.2  | ±    214.81 μs   |     -67.0%