@thoughtpolice 2016-03-29
I've pulled the `INLINE`s and spelling fix. I'm going to try and merge the actual benchmarks with the ones from @michalt in #37, so I'm going to do some refactoring.